Implements brightness factory

+++ includes/message/message.h
@@ -9,7 +9,7 @@ private:
     uint8_t byte_count;
     uint8_t command;
     std::optional<uint8_t> _register;
-    std::optional<uint8_t> data;
+    std::optional<uint16_t> data;
 
 public:
     Message(uint8_t frame_header_h,
@@ -17,7 +17,7 @@ public:
             uint8_t byte_count,
             uint8_t command,
             std::optional<uint8_t> _register,
-            std::optional<uint8_t> data);
+            std::optional<uint16_t> data);
     ~Message();
     uint8_t get_frame_header_h();
     uint8_t get_frame_header_l();
