+++ src/message/message.cpp
@@ -1,4 +1,4 @@
-#include "../includes/message.h"
+#include "../includes/message/message.h"
 #include <iostream>
 
 Message::Message(uint8_t frame_header_h,
@@ -6,7 +6,7 @@ Message::Message(uint8_t frame_header_h,
                  uint8_t byte_count,
                  uint8_t command,
                  std::optional<uint8_t> _register,
-                 std::optional<uint8_t> data)
+                 std::optional<uint16_t> data)
 {
     this->frame_header_h = frame_header_h;
     this->frame_header_l = frame_header_l;
@@ -32,6 +32,7 @@ uint8_t Message::get_command(){
     return this->command;
 }
 
-void Message::execute(){
+void Message::execute()
+{
     
 }
